#1288 Moved ExtractDriverCreator to domain/common. Updated to use ImportInterface.
